requirements.txt
transformers
torch
python-telegram-bot
aiohttp
numpy